backend/requirements.txt
fastapi
uvicorn
requests
pydantic
groq
python-dotenv
numpy
sentence-transformers
PyMuPDF
faiss-cpu
